requirements.txt
open_clip_torch
clip
Pillow
PyYAML
safetensors
torch
tqdm
bayesian-optimization
matplotlib
seaborn
hyperopt
pytest
transformers
numpy
hydra-core
sd-meh==0.9.5
lightgbm 
scikit-learn
openai-clip
tensordict
timm
fairscale
